requirements.txt
# app/requirements.txt
fastapi
discord.py
requests
uvicorn
aiohttp